+++ src/main/kotlin/com/hibob/academy/dao/PetsDao.kt
@@ -5,6 +5,9 @@ import com.hibob.academy.utils.asIs
 import org.jooq.DSLContext
 import org.springframework.stereotype.Component
 import com.hibob.academy.utils.from
+import org.jooq.Record
+import org.jooq.RecordMapper
+import java.time.LocalDate
 
 enum class PetType(val type: String) {
     DOG("dog"),
@@ -13,16 +16,24 @@ enum class PetType(val type: String) {
     FISH("fish");
 }
 
-
 @Component
 class PetsDao(private val sql: DSLContext) {
 
     private val table = PetsTable.instance
 
+    private val ownerMapper = RecordMapper<Record, Pet> {
+            record ->
+        Pet(record[table.id],
+            record[table.name],
+            record[table.type],
+            record[table.companyId],
+            record[table.dateOfArrival]?.time ?: 0L )
+    }
+
     fun petsByType(type: PetType) : List<Pet> {
-        return sql.select(table.name, table.date_of_arrival, table.companyId)
+        return sql.select(table.name, table.dateOfArrival, table.companyId)
             .from(table)
             .where(table.type.eq(type.type))
-            .fetch().map { record3 -> mapToPet(record3) }
+            .fetch(ownerMapper)
     }
 }
